Rename the `/server` exports to `/ssr`
`@convex-dev/auth/server` and `@convex-dev/auth/nextjs/server` hold the
server-side rendering building blocks: the cookie store, the access-token
inspection, the refresh/sign-out handlers, and the Next.js proxy. The
`/server` name reads as "the Convex backend entry point", which is what it
meant in v1 and is not what these modules are.

Rename the entry points to `/ssr` and `/nextjs/ssr`, and move the sources to
`src/ssr/` and `src/nextjs/ssr.tsx` to match.
