+++ examples/nextjs/src/lib/convexAuth.tsx
@@ -1,4 +1,4 @@
-import { setupConvexAuthNextjs } from "@convex-dev/auth/nextjs/server";
+import { setupConvexAuthNextjs } from "@convex-dev/auth/nextjs/ssr";
 import { api } from "@/convex/_generated/api";
 
 // The Next-specific server helpers: proxy (up-front refresh + redirects),
