+++ packages/core/src/ssr/isAuthenticated.test.ts
