+++ packages/core/src/ssr/isAuthenticated.ts
